Add explanatory comments to tools
